+++ service/upgrade_manager.go
@@ -24,6 +24,7 @@ import (
 	"context"
 	"fmt"
 	"net/url"
+	"reflect"
 	"strings"
 	"sync"
 	"time"
@@ -32,6 +33,7 @@ import (
 	"github.com/rs/zerolog"
 	"github.com/ryanuber/columnize"
 
+	"github.com/arangodb-helper/go-helper/pkg/arangod/agency/election"
 	"github.com/arangodb/go-driver"
 	"github.com/arangodb/go-driver/agency"
 	upgraderules "github.com/arangodb/go-upgrade-rules"
@@ -70,9 +72,6 @@ type UpgradeManager interface {
 
 	// RunWatchUpgradePlan keeps watching the upgrade plan until the given context is canceled.
 	RunWatchUpgradePlan(context.Context)
-
-	// UpgradePlanChangedCallback is an agency callback to notify about changes in the upgrade plan
-	UpgradePlanChangedCallback()
 }
 
 // UpgradeManagerContext holds methods used by the upgrade manager to control its context.
@@ -150,6 +149,11 @@ func (p *UpgradePlan) ResetFailures() {
 		e.Failures = 0
 		e.Reason = ""
 	}
+}
+
+// Equals returns true if other plan is the same
+func (p UpgradePlan) Equals(other UpgradePlan) bool {
+	return reflect.DeepEqual(p, other)
 }
 
 // UpgradeEntryType is a strongly typed upgrade plan item
@@ -297,14 +301,14 @@ func (m *upgradeManager) StartDatabaseUpgrade(ctx context.Context, forceMinorUpg
 		return maskAny(err)
 	}
 	m.log.Debug().Msg("Creating lock")
-	lock, err := agency.NewLock(m, api, upgradeManagerLockKey, "", upgradeManagerLockTTL)
+	lock, err := election.NewLock(m, upgradeManagerLockKey, "", upgradeManagerLockTTL)
 	if err != nil {
 		return maskAny(err)
 	}
 
 	// Claim the upgrade lock
 	m.log.Debug().Msg("Locking lock")
-	if err := lock.Lock(ctx); err != nil {
+	if err := lock.Lock(ctx, api); err != nil {
 		m.log.Debug().Err(err).Msg("Lock failed")
 		return maskAny(err)
 	}
@@ -312,7 +316,7 @@ func (m *upgradeManager) StartDatabaseUpgrade(ctx context.Context, forceMinorUpg
 	// Close agency lock when we're done
 	defer func() {
 		m.log.Debug().Msg("Unlocking lock")
-		lock.Unlock(context.Background())
+		lock.Unlock(context.Background(), api)
 	}()
 
 	m.log.Debug().Msg("Reading upgrade plan...")
@@ -540,14 +544,14 @@ func (m *upgradeManager) AbortDatabaseUpgrade(ctx context.Context) error {
 		return maskAny(err)
 	}
 	m.log.Debug().Msg("Creating lock")
-	lock, err := agency.NewLock(m, api, upgradeManagerLockKey, "", upgradeManagerLockTTL)
+	lock, err := election.NewLock(m, upgradeManagerLockKey, "", upgradeManagerLockTTL)
 	if err != nil {
 		return maskAny(err)
 	}
 
 	// Claim the upgrade lock
 	m.log.Debug().Msg("Locking lock")
-	if err := lock.Lock(ctx); err != nil {
+	if err := lock.Lock(ctx, api); err != nil {
 		m.log.Debug().Err(err).Msg("Lock failed")
 		return maskAny(err)
 	}
@@ -555,7 +559,7 @@ func (m *upgradeManager) AbortDatabaseUpgrade(ctx context.Context) error {
 	// Close agency lock when we're done
 	defer func() {
 		m.log.Debug().Msg("Unlocking lock")
-		lock.Unlock(context.Background())
+		lock.Unlock(context.Background(), api)
 	}()
 
 	// Check plan
@@ -820,39 +824,74 @@ func (m *upgradeManager) removeUpgradePlan(ctx context.Context) error {
 	return nil
 }
 
+// waitForPlanChanges returns unbuffered channel on which updated UpgradePlan will be delivered.
+// Closing channel means that there will be no more changes.
+func (m *upgradeManager) waitForPlanChanges(ctx context.Context) chan *UpgradePlan {
+	ch := make(chan *UpgradePlan)
+
+	go func() {
+		defer close(ch)
+		var oldPlan UpgradePlan
+		for {
+			delay := time.Second * 3
+			plan, err := m.readUpgradePlan(ctx)
+			if agency.IsKeyNotFound(err) || plan.IsEmpty() {
+				// Just try later
+			} else if err != nil {
+				// Failed to read plan
+				m.log.Info().Err(err).Msg("Failed to read upgrade plan")
+			} else if !oldPlan.Equals(plan) {
+				ch <- &plan
+				oldPlan = plan
+				delay = time.Second
+			}
+
+			timer := time.NewTimer(delay)
+			select {
+			case <-timer.C:
+				// Continue
+			case <-ctx.Done():
+				// Context canceled
+				if !timer.Stop() {
+					<-timer.C
+				}
+				return
+			}
+		}
+	}()
+
+	return ch
+}
+
 // RunWatchUpgradePlan keeps watching the upgrade plan in the agency.
 // Once it detects that this starter has to act, it does.
 func (m *upgradeManager) RunWatchUpgradePlan(ctx context.Context) {
-	_, myPeer, mode := m.upgradeManagerContext.ClusterConfig()
-	ownURL := myPeer.CreateStarterURL("/")
+	_, _, mode := m.upgradeManagerContext.ClusterConfig()
 	if !mode.HasAgency() {
 		// Nothing to do here without an agency
 		return
 	}
-	registeredCallback := false
-	defer func() {
-		if registeredCallback {
-			m.unregisterUpgradePlanChangedCallback(ctx, ownURL)
-		}
-	}()
+
+	planChanges := m.waitForPlanChanges(ctx)
 	for {
-		delay := time.Minute
-		if !registeredCallback {
-			m.log.Debug().Msg("Registering upgrade plan changed callback...")
-			if err := m.registerUpgradePlanChangedCallback(ctx, ownURL); err != nil {
-				m.log.Info().Err(err).Msg("Failed to register upgrade plan changed callback")
-			} else {
-				registeredCallback = true
-			}
+		var newPlan *UpgradePlan
+		select {
+		case newPlan = <-planChanges:
+			// Plan changed
+		case <-ctx.Done():
+			// Context canceled
+			m.log.Info().Msg("Stopping watching for plan changes: context canceled")
+			return
 		}
-		plan, err := m.readUpgradePlan(ctx)
-		if agency.IsKeyNotFound(err) || plan.IsEmpty() {
-			// Just try later
-		} else if err != nil {
-			// Failed to read plan
-			m.log.Info().Err(err).Msg("Failed to read upgrade plan")
-		} else if plan.IsReady() {
-			// Plan entries have aal been processes
+		if newPlan == nil {
+			// channel was closed
+			m.log.Info().Msg("Stopping watching for plan changes")
+			return
+		}
+
+		plan := *newPlan
+		if plan.IsReady() {
+			// Plan entries have been processed
 			if !plan.Finished {
 				// Let's show the user that we're done
 				if err := m.finishUpgradePlan(ctx, plan); err != nil {
@@ -866,64 +905,8 @@ func (m *upgradeManager) RunWatchUpgradePlan(ctx context.Context) {
 			if err := m.processUpgradePlan(ctx, plan); err != nil {
 				m.log.Error().Err(err).Msg("Failed to process upgrade plan entry")
 			}
-			delay = time.Second
-		}
-
-		select {
-		case <-time.After(delay):
-			// Continue
-		case <-m.cbTrigger.Done():
-			// Continue
-		case <-ctx.Done():
-			// Context canceled
-			return
 		}
 	}
-}
-
-// UpgradePlanChangedCallback is an agency callback to notify about changes in the upgrade plan
-func (m *upgradeManager) UpgradePlanChangedCallback() {
-	m.cbTrigger.Trigger()
-}
-
-// registerUpgradePlanChangedCallback registers our callback URL with the agency
-func (m *upgradeManager) registerUpgradePlanChangedCallback(ctx context.Context, ownURL string) error {
-	// Get api client
-	api, err := m.createAgencyAPI()
-	if err != nil {
-		return maskAny(err)
-	}
-	// Register callback
-	cbURL, err := getURLWithPath(ownURL, "/cb/upgradePlanChanged")
-	if err != nil {
-		return maskAny(err)
-	}
-	ctx, cancel := context.WithTimeout(ctx, time.Second*10)
-	defer cancel()
-	if err := api.RegisterChangeCallback(ctx, upgradePlanKey, cbURL); err != nil {
-		return maskAny(err)
-	}
-	return nil
-}
-
-// unregisterUpgradePlanChangedCallback removes our callback URL from the agency
-func (m *upgradeManager) unregisterUpgradePlanChangedCallback(ctx context.Context, ownURL string) error {
-	// Get api client
-	api, err := m.createAgencyAPI()
-	if err != nil {
-		return maskAny(err)
-	}
-	// Register callback
-	cbURL, err := getURLWithPath(ownURL, "/cb/upgradePlanChanged")
-	if err != nil {
-		return maskAny(err)
-	}
-	ctx, cancel := context.WithTimeout(ctx, time.Second*10)
-	defer cancel()
-	if err := api.UnregisterChangeCallback(ctx, upgradePlanKey, cbURL); err != nil {
-		return maskAny(err)
-	}
-	return nil
 }
 
 // processUpgradePlan inspects the first entry of the given plan and acts upon